+++ NEWS.md
@@ -1,4 +1,13 @@
 
+# sociome 1.2.0
+
+## Major updates:
+- Adapted to tidyr's breaking changes
+
+## Minor updates:
+- Added progress messages for the get_adi() calls that require many calls to tidycensus functions.
+- Code improvements.
+
 # sociome 1.1.0
 
 ## Major updates: